Splitted MatchTimer in two files
